AppContext.loadSpreadsheetMetadataAndPushPreviousIfFails
- Closes https://github.com/mP1/walkingkooka-spreadsheet-dominokit/issues/2560
- Loading unknown spreadsheet id when no previous history should display "/" not remain stuck on /bad-spreadsheet-id/
